# coalgov-ai/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoalGov AI API",
    "description": "Mine governance pilot: inspections, compliance, production approvals, grievances and offline field reporting.",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/cases": {
      "get": {
        "tags": [
          "Cases and offline sync"
        ],
        "summary": "List Cases",
        "operationId": "list_cases_api_cases_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mine_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mine Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 250,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cases and offline sync"
        ],
        "summary": "Add Case",
        "operationId": "add_case_api_cases_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}": {
      "get": {
        "tags": [
          "Cases and offline sync"
        ],
        "summary": "Detail",
        "operationId": "detail_api_cases__case_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/transition": {
      "post": {
        "tags": [
          "Cases and offline sync"
        ],
        "summary": "Change Case",
        "operationId": "change_case_api_cases__case_id__transition_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseTransition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync": {
      "post": {
        "tags": [
          "Cases and offline sync"
        ],
        "summary": "Synchronize",
        "operationId": "synchronize_api_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/documents": {
      "get": {
        "tags": [
          "Documents and evidence"
        ],
        "summary": "Documents",
        "operationId": "documents_api_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Documents and evidence"
        ],
        "summary": "Upload",
        "operationId": "upload_api_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/documents/{document_id}": {
      "get": {
        "tags": [
          "Documents and evidence"
        ],
        "summary": "Detail",
        "operationId": "detail_api_documents__document_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{document_id}/download": {
      "get": {
        "tags": [
          "Documents and evidence"
        ],
        "summary": "Download",
        "operationId": "download_api_documents__document_id__download_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_api_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "mine_id": {
            "type": "string",
            "title": "Mine Id"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "upload_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Id"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "observation",
              "corrective_action",
              "verification",
              "compliance"
            ],
            "title": "Purpose",
            "default": "compliance"
          }
        },
        "type": "object",
        "required": [
          "file",
          "mine_id"
        ],
        "title": "Body_upload_api_documents_post"
      },
      "CaseCreate": {
        "properties": {
          "mine_id": {
            "type": "string",
            "title": "Mine Id"
          },
          "inspection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspection Id"
          },
          "title": {
            "type": "string",
            "maxLength": 220,
            "minLength": 5,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 10,
            "title": "Description"
          },
          "category": {
            "type": "string",
            "enum": [
              "Ventilation",
              "Electrical",
              "Environment",
              "Equipment",
              "Workforce",
              "Fire",
              "Transport",
              "Other"
            ],
            "title": "Category"
          },
          "kind": {
            "type": "string",
            "enum": [
              "observation",
              "violation",
              "incident"
            ],
            "title": "Kind",
            "default": "observation"
          },
          "severity": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Severity"
          },
          "likelihood": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Likelihood"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mine_id",
          "title",
          "description",
          "category",
          "severity",
          "likelihood",
          "captured_at"
        ],
        "title": "CaseCreate"
      },
      "CaseTransition": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "action": {
            "type": "string",
            "enum": [
              "assign",
              "start",
              "submit_action",
              "verify",
              "reject",
              "close",
              "reopen"
            ],
            "title": "Action"
          },
          "assigned_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "note": {
            "type": "string",
            "maxLength": 10000,
            "title": "Note",
            "default": ""
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "action"
        ],
        "title": "CaseTransition"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SyncBatch": {
        "properties": {
          "operations": {
            "items": {
              "$ref": "#/components/schemas/SyncOperation"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Operations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operations"
        ],
        "title": "SyncBatch"
      },
      "SyncOperation": {
        "properties": {
          "operation_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 10,
            "title": "Operation Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "create_case",
              "transition_case",
              "update_inspection",
              "create_production",
              "update_production",
              "create_grievance",
              "transition_grievance"
            ],
            "title": "Kind"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id",
          "kind",
          "payload"
        ],
        "title": "SyncOperation"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}